fixes initializing Span<PrimitiveTypeBiggerThanOneByte> with stackalloc PrimitiveTypeBiggerThanOneByte [const]
